\name{UnconfoundedBIMA-package}
\alias{UnconfoundedBIMA-package}
\alias{UnconfoundedBIMA}
\docType{package}
\title{
\packageTitle{UnconfoundedBIMA}
}
\description{
\packageDescription{UnconfoundedBIMA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UnconfoundedBIMA}
\packageIndices{UnconfoundedBIMA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{UnconfoundedBIMA}

Maintainer: \packageMaintainer{UnconfoundedBIMA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
